added hospital_3F scene and renamed PlayerMovement.cs script
