+++ Assets/Prefabs/Door.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3cd05ab6de143fd499488ac5ebbff0a5
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
