+++ Assets/Prefabs/Small Wall.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bdecb4db3f3f0cd40a7d4436d9198030
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
